+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/IsoMdocFallbackCredentialScheme.kt
@@ -1,6 +1,5 @@
-package at.asitplus.wallet.fallbackCredential.isoMdocFallbackCredentialScheme
+package at.asitplus.wallet.lib.data
 
-import at.asitplus.wallet.lib.data.ConstantIndex
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.ISO_MDOC
